move gasdev() into a separate code module.
